Osa 4 tehty

+++ osa4/controllers/blogs.js
@@ -1,14 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { tokenExtractor, userExtractor } = require('../utils/middleware')
 
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', tokenExtractor, userExtractor, async (request, response) => {
     const body = request.body
+    const user = request.user
 
     if (!body.title || !body.author || !body.url) {
         return response.status(400).json({
@@ -21,9 +24,12 @@ blogsRouter.post('/', async (request, response) => {
         author: body.author,
         url: body.url,
         likes: body.likes || 0,
+        user: user._id,
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
 
 })
@@ -37,7 +43,15 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', tokenExtractor, userExtractor, async (request, response) => {
+
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() !== user._id.toString() ) {
+        return response.status(401).json({ error: 'unauthorized user token' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
